[DEV] `#populate()` now checks if the properties declared in `__children` are also present in the "parent" model-description and throws and `Error` if not.
In order to be able to do this check recursively and for the sake of sane code structure, a "computedProperty" now gets passed the full `Molder` instance in addition to the `preparedOrigin` object and the `#extractNamespace()` method was made `public` (removed the underscore ('_') prefix as well as removed the `@private` tag from the documentation). So, a "computedProperty" is now able to use the full toolkit of a `json-molder` instance.